+++ SeriousMimesUnityProject/Assets/Prefabs/Glow.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fb372ed128a1f6e408b1e1b10eb2cbde
+timeCreated: 1485060568
+licenseType: Pro
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
