+++ SeriousMimesUnityProject/Assets/Scripts/CollisionBuddy.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1e76bbe75a7073547a9dabd68c653606
+timeCreated: 1485062082
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
