+++ SeriousMimesUnityProject/Temp/UnityTempFile-2148288e56aec884fbb5ce1f410bdb85
@@ -1,0 +1,144 @@
+-debug
+-target:library
+-nowarn:0169
+-out:Temp/Assembly-CSharp.dll
+-r:"D:/Program Files/Unity/Editor/Data/Managed/UnityEngine.dll"
+-r:Library/ScriptAssemblies/Assembly-CSharp-firstpass.dll
+-r:"D:/Program Files/Unity/Editor/Data/UnityExtensions/Unity/GUISystem/UnityEngine.UI.dll"
+-r:"D:/Program Files/Unity/Editor/Data/UnityExtensions/Unity/Networking/UnityEngine.Networking.dll"
+-r:"D:/Program Files/Unity/Editor/Data/Managed/UnityEditor.dll"
+-r:"D:\Program Files\Unity\Editor\Data\PlaybackEngines\AppleTVSupport\UnityEditor.iOS.Extensions.Common.dll"
+-r:"D:\Program Files\Unity\Editor\Data\PlaybackEngines\AppleTVSupport\UnityEditor.iOS.Extensions.Xcode.dll"
+-r:"D:/Program Files/Unity/Editor/Data/PlaybackEngines/iOSSupport\UnityEditor.iOS.Extensions.Xcode.dll"
+-r:"D:/Program Files/Unity/Editor/Data/PlaybackEngines/iOSSupport\UnityEditor.iOS.Extensions.Common.dll"
+-define:UNITY_5_3_OR_NEWER
+-define:UNITY_5_4_OR_NEWER
+-define:UNITY_5_4_0
+-define:UNITY_5_4
+-define:UNITY_5
+-define:ENABLE_NEW_BUGREPORTER
+-define:ENABLE_AUDIO
+-define:ENABLE_CACHING
+-define:ENABLE_CLOTH
+-define:ENABLE_DUCK_TYPING
+-define:ENABLE_EDITOR_RETINA
+-define:ENABLE_RETINA_GUISTYLES
+-define:ENABLE_FRAME_DEBUGGER
+-define:ENABLE_GENERICS
+-define:ENABLE_HOME_SCREEN
+-define:ENABLE_IMAGEEFFECTS
+-define:ENABLE_LIGHT_PROBES_LEGACY
+-define:ENABLE_MICROPHONE
+-define:ENABLE_MULTIPLE_DISPLAYS
+-define:ENABLE_PHYSICS
+-define:ENABLE_PLUGIN_INSPECTOR
+-define:ENABLE_SHADOWS
+-define:ENABLE_SPRITERENDERER_FLIPPING
+-define:ENABLE_SPRITES
+-define:ENABLE_SPRITE_POLYGON
+-define:ENABLE_TERRAIN
+-define:ENABLE_RAKNET
+-define:ENABLE_UNET
+-define:ENABLE_UNITYEVENTS
+-define:ENABLE_VR
+-define:ENABLE_WEBCAM
+-define:ENABLE_WWW
+-define:ENABLE_CLOUD_SERVICES
+-define:ENABLE_CLOUD_SERVICES_COLLAB
+-define:ENABLE_CLOUD_SERVICES_ADS
+-define:ENABLE_CLOUD_HUB
+-define:ENABLE_CLOUD_PROJECT_ID
+-define:ENABLE_CLOUD_SERVICES_CRASH_REPORTING
+-define:ENABLE_CLOUD_SERVICES_PURCHASING
+-define:ENABLE_CLOUD_SERVICES_ANALYTICS
+-define:ENABLE_CLOUD_SERVICES_UNET
+-define:ENABLE_CLOUD_SERVICES_BUILD
+-define:ENABLE_CLOUD_LICENSE
+-define:ENABLE_EDITOR_METRICS
+-define:ENABLE_EDITOR_METRICS_CACHING
+-define:INCLUDE_DYNAMIC_GI
+-define:INCLUDE_GI
+-define:INCLUDE_IL2CPP
+-define:INCLUDE_DIRECTX12
+-define:PLATFORM_SUPPORTS_MONO
+-define:RENDER_SOFTWARE_CURSOR
+-define:INCLUDE_PUBNUB
+-define:ENABLE_LOCALIZATION
+-define:ENABLE_ANDROID_ATLAS_ETC1_COMPRESSION
+-define:ENABLE_EDITOR_TESTS_RUNNER
+-define:UNITY_STANDALONE_WIN
+-define:UNITY_STANDALONE
+-define:ENABLE_SUBSTANCE
+-define:ENABLE_TEXTUREID_MAP
+-define:ENABLE_RUNTIME_GI
+-define:ENABLE_MOVIES
+-define:ENABLE_NETWORK
+-define:ENABLE_CRUNCH_TEXTURE_COMPRESSION
+-define:ENABLE_UNITYWEBREQUEST
+-define:ENABLE_EVENT_QUEUE
+-define:ENABLE_CLUSTERINPUT
+-define:ENABLE_WEBSOCKET_HOST
+-define:ENABLE_MONO
+-define:ENABLE_PROFILER
+-define:DEBUG
+-define:TRACE
+-define:UNITY_ASSERTIONS
+-define:UNITY_EDITOR
+-define:UNITY_EDITOR_64
+-define:UNITY_EDITOR_WIN
+-define:UNITY_TEAM_LICENSE
+-define:UNITY_PRO_LICENSE
+-define:CROSS_PLATFORM_INPUT
+Assets/Beautify/Demos/DemoSources/Scripts/CameraAnimator.cs
+Assets/Beautify/Demos/DemoSources/Scripts/Demo1.cs
+Assets/Beautify/Demos/DemoSources/Scripts/Demo2.cs
+Assets/Beautify/Demos/DemoSources/Scripts/FreeCameraMove.cs
+Assets/Beautify/Scripts/Beautify.cs
+"Assets/Realistic Effects Pack/Materials/Objects/FireJetPack/FlyDemo.cs"
+"Assets/Realistic Effects Pack/Scripts/Demo/DemoGUI.cs"
+"Assets/Realistic Effects Pack/Scripts/Demo/MyGUI3_1.cs"
+"Assets/Realistic Effects Pack/Scripts/Demo/RealtimeReflection.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/AddMaterialOnHit.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/AnimatorBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/IceOffsetBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/LazyLoad.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/LineProjectileCollisionBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/ProjectileCollisionBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Balls/ScaleParticlesFromBound.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Buffs/AnimationSpeedDebuff.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Buffs/DebuffOnEnemyFromCollision.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Buffs/ShieldCollisionBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Buffs/WaterUvAnimation.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Objects/MoveOnGround.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Objects/OffsetOnNormal.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Objects/OnRigidbodySendCollision.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Objects/ResetPositionOnDiactivated.cs"
+"Assets/Realistic Effects Pack/Scripts/Prefabs/Shots/LineRendererBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/Billboard.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/CollisionActiveBehaviour.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/DeadTime.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/DistortionMobileCamera.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/EffectSettings.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/EffectsColorizer.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/EffectsParticleSystemScaler.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutLight.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutParticles.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutScale.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutShaderColor.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutShaderFloat.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutSound.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FadeInOutStatus.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/FixShaderQueue.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/OnStartSendCollision.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/RotateAround.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/SetPositionOnHit.cs"
+"Assets/Realistic Effects Pack/Scripts/Share/UVTextureAnimator.cs"
+Assets/Scripts/Beam.cs
+Assets/Scripts/CameraFollowGuy.cs
+Assets/Scripts/CustomCharacterController.cs
+Assets/Scripts/EnemyBehaviour.cs
+Assets/Scripts/EnemySpawner.cs
+Assets/Scripts/Patrol.cs
+Assets/Scripts/WaveGun.cs
+-r:"D:\Program Files\Unity\Editor\Data\Mono\lib\mono\unity\System.Runtime.Serialization.dll"
+-r:"D:\Program Files\Unity\Editor\Data\Mono\lib\mono\unity\System.Xml.Linq.dll"
